requirements.txt
flask
transformers
torch
accelerate
python-dotenv